second commit

+++ src/main/java/com/sarahi/springdata/shoppingcart/repos/OrderHistoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sarahi.springdata.shoppingcart.model.OrderHistory;
 
-public interface OrderHistoryRepository extends JpaRepository<OrderHistory, Integer> {
+public interface OrderHistoryRepository extends JpaRepository<OrderHistory, Long> {
 
 }
