+++ src/main/java/com/sarahi/springdata/shoppingcart/repos/ProductRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.sarahi.springdata.shoppingcart.model.Product;
 
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
-	Product findById(int product_id);
+	Product findById(long product_id);
 	Product findByName(String name);
 	List<Product> findByPrice(int price);
 }
