+++ src/main/java/com/sarahi/springdata/shoppingcart/repos/UserRepository.java
@@ -6,10 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sarahi.springdata.shoppingcart.model.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 	List<User> findByName(String name);
 	User findByEmail(String email);
-	User findById(int user_id);
+	User findById(long user_id);
 
 }
